Fix closing entries by pasting.  Improve later.
The previous approach of touching functional fields was too slow: the
SQL approach takes about a minute whereas the former was interrupted
after an hour.

The only way to do this extensible is by changing the core OpenERP
accounting module to use a SQL representation that could be changed
elsewhere but that would require much coordination for this module to
wait for it.