Improve performance for big analytic accounts.
When posting a journal entry many of it's line may affect the same analytic
account.  In such cases the `create_analytic_lines` drops the analytic lines
of the move and creates new ones.

For each, analytic line created Odoo triggers the computation of the fields
'invoiced', 'self_balance', etc, which iterate over the entire set of analytic
lines (which is yet incomplete).  Doing so is a performance headache for
analytic accounts with thousands of lines.

We can wait until all lines have been created to perform the computation only
once.