Ensure invoiced/expended to be updated when removing moves.
When cancelling an invoice it removes it associated journal entry.  This makes
Odoo (via an ``ondelete='cascade'`` in ``account/account_analytic_line.py``)
to remove the analytic lines.  However, the method ``_compute_invoiced()`` is
not being called.

I first thought that it was due to the usage of the context
AVOID_LENGTHY_COMPUTATION, but that I've found to be unrelated.

I still need to confirm that ``@api.depends`` over one2many fields are not
triggered if they're removed via ``ondelete='cascade'``.