Improve performance of the computation of the primary salesperson.
The computation of the primary salesperson is quite expensive in Odoo 10.
This has a heavy impact in the validation of an invoice (even those with a
single line, and which has no impact in salesperson).

This commit improves the computation in two stages:

- When creating analytic lines, call the compute primary salesperson method
  if any invoice is client invoice/refund.

- In the compute method avoid the 'search()' and perform linear first of the
  salesperson in the sales invoices/refunds.

  It seems that the domain was not efficient (perhaps some fields lacked the
  `auto_join` attribute).  Our method is **much** quicker.

As a bonus the check of having more than one salesperson is now removed since
it's not being used (nor it's useful).