Add 'counterpart' accounts in journal items.
This will allow to see the flow from one account to another.

The problem:

I'm trying to create an opening statement so that I can reconcile all
posted (but not reconciled) payments migrated from Odoo 8.

It seems (I'm still munging the code) that each statement line is reconciled
with a payment (which explain the cash flow).  In turn the payment is the way
to reconcile the journal items.

However, I know the amount of money I have in a given bank/cash account.  But
in order to create a *cash flow* that leads to this amount I have to go from 0
to the current amount with three types of movements:

- Client payments (collecting money to increase the account's balance)
- Supplier payments (decrease the account's balance)
- Internal transfers from one cash/bank account to another.

I can have other types (I used to create journal entries moving from
receivables or payables to bank and vice-verse.)

The solution is to collect wholesale information about the journal items that
touch the bank account with regards to other accounts (the counterparts).