Rewrite generation of commissions.
Use the right attribute name for the partner's Payables Account.  In Odoo 8 it
was 'property_account_payable', now is 'property_account_payable_id'.

At the same time we try to use recordsets instead of ids.

Also, the `message_follower_ids` was a relation to partners, now it relates to
'mail.follower' objects.  In any case, it's better to rely on methods to
change data.

Show the created invoice.  This helped us to notice that the invoice was being
created empty.  This is because Odoo is issuing a UNLINKALL command and then
only UPDATE_RELATED (which does not make the link).

Fixes mercurio/xhg#61.