Split the unbalanced-keep track from proper currency.
Commit 5e46d154d53613ceaedc3652a68120fa8a9c8175 introduced a track of how
much *unbalanced* a move was.  It introduced the change in the addon
`xopgi_proper_currency`.  However, this feature is only partially related to
proper currency.  It can be activated without needing to split the proper
debit/credit columns.

We extracted the core of the unbalanced amount to 'xopgi_account'.

To avoid having to create an integration addon (i9n), now
`xopgi_proper_currency` depends on `xopgi_account`.

Also the name of 'balance' was inaccurate, it's really a measure of how
much **unbalaced** a move is.