Remove context AVOID_LENGTHY_COMPUTATION.
Ensures the primary salesperson is properly computed.  Solves issue
https://odoo.merchise.org/web#id=1135&view_type=form&model=project.issue&action=480&active_id=28

When create analytic lines for supplier invoices having more than a line we
didn't set the primary salesperson:

- We wrapped the computation within AVOID_LENGTHY_COMPUTATION
- And we didn't call the compute_primary_salesperson for non-income move line.

It seems Odoo 10 requires you to set the computed field, otherwise it blanks
it.

The AVOID_LENGTHY_COMPUTATION was used to avoid Odoo 8's overly computation of
analytic lines.  In Odoo 10, they only create the analytic lines when posting
the journal entry.  So, this hack is no longer needed.