Introduce the normal-balance of account to avoid over-matching.
If an account with a *debit* normal balance had more credit than debit, we
were taking the 'abs' of the balance to create the match.  But this should not
be allowed.

We introduce a couple of classes for accounts with a:

- debit normal balance, or
- credit normal balance

This classes implement the *core* of computing the balance and increasing or
decreasing the balance of an account.